+++ www/go/modules/community/comments/model/Comment.php
@@ -203,45 +203,46 @@ class Comment extends AclItemEntity {
 	protected function internalSave()
 	{
 
-		if($this->isNew()) {
-			$this->createAlerts();
-		}
+//		if($this->isNew()) {
+//			$this->createAlerts();
+//		}
 		$this->images = Blob::parseFromHtml($this->text);
 		return parent::internalSave();
 	}
-
-	private function createAlerts() {
-		$entity = $this->findEntity();
-		$aclId = $entity->findAclId();
-		if(!$aclId) {
-			return;
-		}
-
-		$excerpt = StringUtil::cutString(strip_tags($this->text), 50);
-
-		$userIds = go()->getDbConnection()->selectSingleValue('userId')
-			->from('core_user_group', 'ug')
-			->join('core_acl_group', 'ag', 'ag.groupId = ug.groupId')
-			->where('ag.aclId', '=', $aclId);
-
-		foreach($userIds as $userId) {
-
-			if($userId == go()->getAuthState()->getUserId()) {
-				continue;
-			}
-
-			$alert = $entity->createAlert(new DateTime(), 'comment', $userId)
-				->setData([
-					'type' => 'comment',
-					'createdBy' => go()->getAuthState()->getUserId(),
-					'excerpt' => $excerpt
-				]);
-
-			if(!$alert->save()) {
-				throw new SaveException($alert);
-			}
-		}
-	}
+//
+//	private function createAlerts() {
+//
+//		$entity = $this->findEntity();
+//		$aclId = $entity->findAclId();
+//		if(!$aclId) {
+//			return;
+//		}
+//
+//		$excerpt = StringUtil::cutString(strip_tags($this->text), 50);
+//
+//		$userIds = go()->getDbConnection()->selectSingleValue('userId')
+//			->from('core_user_group', 'ug')
+//			->join('core_acl_group', 'ag', 'ag.groupId = ug.groupId')
+//			->where('ag.aclId', '=', $aclId);
+//
+//		foreach($userIds as $userId) {
+//
+//			if($userId == go()->getAuthState()->getUserId()) {
+//				continue;
+//			}
+//
+//			$alert = $entity->createAlert(new DateTime(), 'comment', $userId)
+//				->setData([
+//					'type' => 'comment',
+//					'createdBy' => go()->getAuthState()->getUserId(),
+//					'excerpt' => $excerpt
+//				]);
+//
+//			if(!$alert->save()) {
+//				throw new SaveException($alert);
+//			}
+//		}
+//	}
 
 	public function title()
 	{